Update detect/1
Change the `detect/1` to `detect/2` which takes the bus address of the
device that will be investigated. The function will ask the device for
its id and the response will provide the id in the return value. This
provides a programmatic way ensure the right sensor(s) are available for
use before starting the BMP280 server.